More accurate REST/gRPC response status codes for errors
Currently MLServer returns a 400 HTTP response or INVALID_ARGUMENT gRPC response in all error cases. It would be useful to distinguish between different classes of error, in particular indicating the model-not-found case with a 404 or gRPC NOT_FOUND status. Model-mesh uses NOT_FOUND to take corrective action i.e. re-load the model in question if necessary.