test(huggingface): Use Optimum supported task
I've changed tasks as well avoid a `can_generate` error on the default
chosen model for the question-answering task.

```
E               ValueError: Asked to export a gpt2 model for the task question-answering, but the Optimum ONNX exporter only supports the tasks feature-extraction, feature-extraction-with-past, text-generation, text-generation-with-past, text-classification, token-classification for gpt2. Please use a supported task. Please open an issue at https://github.com/huggingface/optimum/issues if you would like the task question-answering to be supported in the ONNX export for gpt2.
```